+++ packages/@expo/cli/src/start/server/metro/createServerRouteMiddleware.ts
@@ -4,10 +4,10 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import { ExpoResponse } from '@expo/server';
-import { createRequestHandler } from '@expo/server/build/vendor/http';
+
 import requireString from 'require-from-string';
 import resolve from 'resolve';
+import resolveFrom from 'resolve-from';
 import { promisify } from 'util';
 
 import { ForwardHtmlError } from './MetroBundlerDevServer';
@@ -15,6 +15,7 @@ import { bundleApiRoute } from './bundleApiRoutes';
 import { fetchManifest } from './fetchRouterManifest';
 import { getErrorOverlayHtmlAsync, logMetroError, logMetroErrorAsync } from './metroErrorInterface';
 import { Log } from '../../../log';
+import { CommandError } from '../../../utils/errors';
 
 const debug = require('debug')('expo:start:server:metro') as typeof console.log;
 
@@ -35,6 +36,16 @@ export function createRouteHandlerMiddleware(
     getStaticPageAsync: (pathname: string) => Promise<{ content: string }>;
   }
 ) {
+  if (!resolveFrom.silent(projectRoot, 'expo-router')) {
+    throw new CommandError(
+      'static and server rendering requires the expo-router package to be installed in your project.'
+    );
+  }
+
+  const { ExpoResponse } = require('expo-router/server') as typeof import('expo-router/server');
+  const { createRequestHandler } =
+    require('@expo/server/build/vendor/http') as typeof import('@expo/server/build/vendor/http');
+
   return createRequestHandler(
     { build: '' },
     {